Functools added, button changes colour accordingly
functiontools added an new way to reduce lines of code.
took out a for loop that wasn't needed and combined it into one.
added diceclicked array to calculate if the dice needs to be rolled or not.
added def click to make it either red or black depending if a roll needs to be calculated